% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_success.R
\name{add_success}
\alias{add_success}
\title{Add Success Stats to a Plays Dataframe}
\usage{
add_success(plays_df)
}
\arguments{
\item{plays_df}{dataframe an existing dataframe object as provided by the
get_plays() function. Ideally this dataframe has not been modified yet.}
}
\description{
Adds success rate metrics to an existing play-by-play dataframe as
provided by the get_plays() function.
}
\examples{
library(statRdaysCFB)
plays_df <- get_plays(start_year = 2021, end_year = 2021)
plays_w_success <- add_success(plays_df)
}
