% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wpa.r
\name{add_wpa}
\alias{add_wpa}
\title{Add win probability added to play-by-play}
\usage{
add_wpa(plays_df)
}
\arguments{
\item{plays_df}{a dataframe, ideally the one provided by the
get_plays() function.}
}
\description{
Add win probability added to play-by-play
}
