% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cfbd_api}
\alias{cfbd_api}
\title{Make a direct call to the cfbd api}
\usage{
cfbd_api(url, key)
}
\arguments{
\item{url}{character - Full url endpoint you want to access.}

\item{key}{character - Your api key}
}
\value{
a tibble
}
\description{
Directly call an API endpoint providing the full url and your key. Note, you
usually don't need to do this as the get_anything() function should be able to
call almost any endpoint for you.
}
