% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_anything}
\alias{get_anything}
\title{Get Anything From the CFBD API}
\usage{
get_anything(
  url,
  start_year = 2021,
  end_year = 2021,
  start_week,
  end_week,
  key
)
}
\arguments{
\item{url}{string. The url of the API endpoint WITHOUT any parameters added onto the end.
If you have any parameters to add, please use the provided fields and we will add them
to the url for you.}
}
\description{
Get data from any API endpoint on cfbdata.com
}
