% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_betting}
\alias{get_betting}
\title{Get betting data from the cfbd api}
\usage{
get_betting(start_year, end_year, start_week, end_week, season_type = "both")
}
\arguments{
\item{start_year}{int. The first season you want betting data for.}

\item{end_year}{int. The last season you want betting data for.}

\item{start_week}{int (optional). The first week of the season you want betting data for.}

\item{end_week}{int (optional). The last week of the season you want betting data for.}

\item{season_type}{character. "regular", "postseason", or "both". Default is "both".}
}
\value{
tibble of game-level betting data
}
\description{
Get the betting data from collegefootballdata.com for a specified range of
years or weeks. Most years contain spread data at a minimum, and recent years
(2021-present) have moneylines and over/unders as well.
}
\details{
In the case that multiple providers have offered lines, there will be multiple entries returned
for each game. We've taken the liberty of taking the average line of the different
providers in cases where there's more than one line given. However, if you'd
prefer to get all the lines from the various providers yourself, you can instead
make a call to to betting endpoint directly (https://api.collegefootballdata.com/lines)
using the get_anything() function.
}
