% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_colors}
\alias{get_colors}
\title{Get Team Colors, Mascots and Abbreviations}
\usage{
get_colors()
}
\description{
Get a list of each school's primary, secondary colors, mascots,
team nicknames, abbreviations, and more from the teams endpoint.
}
