% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_drives}
\alias{get_drives}
\title{Get the drive data from collegefootballdata.com for a specified range of
years or weeks.}
\usage{
get_drives(start_year, end_year, start_week, end_week)
}
\arguments{
\item{start_year}{int. The first season you want drive data for.}

\item{end_year}{int. The last season you want drive data for.}

\item{start_week}{int. The first week of the season you want drive data for.}

\item{end_week}{int. The last week of the season you want drive data for.}
}
\value{
tibble of drive-level data
}
\description{
Get the drive data from collegefootballdata.com for a specified range of
years or weeks.
}
