% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_games}
\alias{get_games}
\title{Get the game data from collegefootballdata.com for a specified range of
years or weeks. Note that you can request future weeks for the current season
and you will essentially get back the schedule for the season. If the game is completed,
you will get the final score along with everything else.}
\usage{
get_games(start_year, end_year, start_week, end_week)
}
\arguments{
\item{start_year}{int. The first season you want game data for.}

\item{end_year}{int. The last season you want game data for.}

\item{start_week}{int. The first week of the season you want game data for.}

\item{end_week}{int. The last week of the season you want game data for.}
}
\value{
tibble of game-level data
}
\description{
Get the game data from collegefootballdata.com for a specified range of
years or weeks. Note that you can request future weeks for the current season
and you will essentially get back the schedule for the season. If the game is completed,
you will get the final score along with everything else.
}
