% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_of_5}
\alias{group_of_5}
\title{List of group of five conference names historically.}
\format{
a character vector
}
\source{
created
}
\usage{
group_of_5
}
\description{
List of group of five conference names historically.
}
\keyword{datasets}
