% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{my_key}
\alias{my_key}
\title{Set key as environment variable and make call to the cfbd api}
\format{
An object of class \code{character} of length 1.
}
\usage{
my_key
}
\description{
Set key as environment variable and make call to the cfbd api
}
\keyword{datasets}
