% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no_action_plays}
\alias{no_action_plays}
\title{List of all play types where no play occurs (timeouts, end of quarters, ...)}
\format{
a character vector
}
\source{
created
}
\usage{
no_action_plays
}
\description{
List of all play types where no play occurs (timeouts, end of quarters, ...)
}
\keyword{datasets}
