% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power_5}
\alias{power_5}
\title{List of power five conference names historically.}
\format{
a character vector
}
\source{
created
}
\usage{
power_5
}
\description{
List of power five conference names historically.
}
\keyword{datasets}
