% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrimmage_plays_all}
\alias{scrimmage_plays_all}
\title{List of all scrimmage play types}
\format{
a character vector:
}
\source{
cfbd pbp data
}
\usage{
scrimmage_plays_all
}
\description{
This contains the play names typically used in the pbp
data for scrimmage plays. This does not include special teams plays
like kickoffs or punts.
}
\keyword{datasets}
