% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrimmage_plays_kicks}
\alias{scrimmage_plays_kicks}
\title{List of all plays that result in a kick}
\format{
a character vector
}
\source{
created
}
\usage{
scrimmage_plays_kicks
}
\description{
List of all plays that result in a kick
}
\keyword{datasets}
