% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrimmage_plays_non_turnover}
\alias{scrimmage_plays_non_turnover}
\title{List of all plays that don't result in a turnover}
\format{
a character vector
}
\source{
created
}
\usage{
scrimmage_plays_non_turnover
}
\description{
List of all plays that don't result in a turnover
}
\keyword{datasets}
