% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrimmage_plays_pass}
\alias{scrimmage_plays_pass}
\title{List of all passing play types}
\format{
a character vector
}
\source{
created
}
\usage{
scrimmage_plays_pass
}
\description{
List of all passing play types
}
\keyword{datasets}
