% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrimmage_plays_rush}
\alias{scrimmage_plays_rush}
\title{List of all rushing play types}
\format{
a character vector
}
\source{
created
}
\usage{
scrimmage_plays_rush
}
\description{
List of all rushing play types
}
\keyword{datasets}
