% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrimmage_plays_turnover}
\alias{scrimmage_plays_turnover}
\title{List of all plays that DO result in a turnover}
\format{
a character vector
}
\source{
created
}
\usage{
scrimmage_plays_turnover
}
\description{
This contains play types such as interceptions and fumbles.
This does NOT include punts.
}
\keyword{datasets}
