% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_cfb_api_key}
\alias{set_cfb_api_key}
\title{Set your personal collegefootballdata.com key to your environment}
\usage{
set_cfb_api_key(key)
}
\arguments{
\item{key}{character. Your personal API key. If you don't have a key, you can register for one for
free at https://collegefootballdata.com/key}
}
\description{
Takes your input and assigns it to the "cfbd_staturdays_key" key
in your .Renviron profile.
Prereq: You need to get an API key, and then locate your .Renviron file and add a new line
cfbd_staturdays_key = "mysecretkey"
}
\details{
This will create a new key in your .Renviron file with the value you assign to it.
You will then access that key using my_key() which will be created for you when calling
most "get" functions. You can also access your key at any time using
Sys.getenv("cfbd_staturdays_key") or change it by using set_cfb_api_key() or directly using
Sys.setenv("cfbd_staturdays_key" = "yourapikeyhere")
}
