% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staturdays_theme.R
\docType{data}
\name{staturdays_col_list}
\alias{staturdays_col_list}
\title{Get Staturdays Colors and Themes to Build Beautiful Visuals}
\format{
An object of class \code{character} of length 8.
}
\usage{
staturdays_col_list
}
\description{
Get Staturdays Colors and Themes to Build Beautiful Visuals
}
\keyword{datasets}
