% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staturdays_theme.R
\name{staturdays_colors}
\alias{staturdays_colors}
\title{Get a list of staturdays colors and hexes}
\usage{
staturdays_colors(...)
}
\description{
Simply call staturdays_colors() to get a list of our
theme colors and their hex values, or provide a color name and it will
return the hex.
}
