% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staturdays_palette}
\alias{staturdays_palette}
\title{List of the staturdays color palette}
\format{
a character vector
}
\usage{
staturdays_palette
}
\description{
List of the staturdays color palette
}
\keyword{datasets}
