% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staturdays_theme.R
\name{staturdays_ramp}
\alias{staturdays_ramp}
\title{Create a color ramp}
\usage{
staturdays_ramp(x)
}
\description{
This can be used for gt() tables if you want
a gradient from one color to another representing low-to-high values.
}
