% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{staturdays_theme}
\alias{staturdays_theme}
\title{ggplot2 theme for use with plotting}
\format{
ggplot2 theme() object.
}
\usage{
staturdays_theme
}
\description{
This sets some useful defaults to make plots look more polished.
}
\details{
To use it, simply put this line at the end of your ggplot code:
\itemize{
\item staturdays_theme
}

Note: This is not a function. Don't use ().
}
\keyword{datasets}
