% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staturdays_theme.R
\docType{data}
\name{staturdays_theme}
\alias{staturdays_theme}
\title{Create ggplot2 Objects With Staturdays Theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 11.
}
\usage{
staturdays_theme
}
\description{
Create ggplot2 Objects With Staturdays Theme
}
\keyword{datasets}
